ATO-1868: Rename KeyPairHelper to KeyPairUtils
- Now we have removed the unnecessary caching in this class, it is stateless, so we can rename it to KeyPairUtils and move it into the utils package.
